pyproject.toml
[tool.poetry]
authors = ["CircuitSacul <circuitsacul@icloud.com>", "Lunarmagpie <Bambolambo0@gmail.com>", "trag1c <trag1cdev@yahoo.com>"]
description = "The Docket Bot"
license = "MIT"
name = "docket"
version = "0.1.0"

[tool.poetry.dependencies]
hikari = "^2.0.0-alpha.108"
hikari-crescent = "^0.1.0"
python = ">=3.8,<3.11"
lupa = "^1.13"
apgorm = "^1.0.0-beta.11"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.961"
nox = "^2022.1.7"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
line-length=88
skip-magic-trailing-comma=true
target-version=['py38']

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
namespace_packages = true
strict = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
  "lupa.*",
  "asyncpg.*"
]

[tool.isort]
profile = "black"
line_length = 88

